Add explicit dependency on JaxB
Java 9 added the Jigsaw module system and moved `java.xml.bind` into
a module. This module is not active by default. It can be be activated
by running with `--add-modules java.xml.bind` however as it is scheduled
to be removed in Java 10 it is prudent to migrate away from it now.

As such the `jaxb-api`, glassfishes `jaxb-runtime` reference
implementation and its dependency `javax.activation` have been added as
maven dependencies.

Additionally the `maven-plugin-plugin` failed to execute the help goal.
This appeared to be caused by the initializer blocks of the anonymous
subclass of the array list in BinaryType.

This has been resolved by incrementing the plugin version and using
`Arrays.asList` and replacing the use of `ArrayList` with the more
general `List` interface where necessary.